reverse-eng/requirements.txt
pyserial
#minimalmodbus
#pymodbus
